Add Flask routes for Spotify OAuth, setup html/css
In main.py, '/' route checks for a valid token. If the token is invalid
or expired, return to login. If the token is valid, redirect to search
page. The '/callback' route handles OAuth redirect and obtains the
access token needed to make requests on the user's behalf. The '/search'
route renders the search page template. The '/logout' route redirects to
home. Basic html boilerplate was added to index.html, which is linked to
styles.css.

API wrappers used:
- Spotify web API: Spotipy
- TMDb API: tmdbsimple