Add basic searching function and popup dialog
The '/search' endpoint displays index.html, which contains a search bar.
In script.js, an event listener attached to the input element updates
results while the user is typing. The page is cleared if there is no
input. If there is, the event listener fetches data from the
'/search-api' endpoint and dynamically renders the cards on the page.
The search_api() function uses search_movies() in tmdb.py, which uses
the tmdbsimple api wrapper to fetch relevant movies based on a query.
Another event listener attached to the cards calls openModal(), which
displays popup dialog on the page that can be cleared. This will be used
to give a preview of the created playlist, which can be added to the
user's personal playlists with a button.

Want to improve searching function
- clear previous search results rendered while typing so that most
relevant searches appear near the top